+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unrealesed
+
+* feat: app typography
+
 ## 0.0.0+2
 
 * chore: tune analysis options
@@ -9,7 +13,6 @@
 * chore: automate PR author auto assign
 * chore: setup VSCode spellchecker project dictionary
 * chore: tune build app workflow
-* feat: app typography
 
 ## 0.0.0+1